fix(viewer): surface Lists panel run/import failures instead of console.error-ing them
`ListPanel.handleExecuteList` and `handleImport` caught a failure and only
logged it — the spinner stopped, the panel stayed on its current view, and
nothing on screen said why. That reads identically to a genuine empty
result. It became user-reachable once `compileNameMatcher`
(packages/lists/src/name-pattern.ts, #4262/#4292, stacked here) started
THROWING on a ReDoS-guard-rejected `/regex/` name-pattern column instead of
falling back silently.

Adds a `listError` store field (mirroring SearchModal.filter.tsx's
`searchFilterError`) and a dismissible "List failed" banner
(`ListErrorBox.tsx`, styled like the advanced filter's `FilterErrorBox`),
rendered above whichever view is active. A genuinely empty result still
clears the error and publishes an ordinary (possibly all-null) result —
only a caught exception sets it.

Also splits `ListPanel.tsx`'s List Library / List Item sub-components into
`ListLibrary.tsx`, to keep the file under the module-size budget after the
error-surfacing addition (the allowlist row is now stale and removed).

Closes #4317
